backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
gunicorn==21.2.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
pydantic==2.5.0
pydantic-settings==2.1.0
python-dotenv==1.0.0
python-jose==3.3.0
passlib==1.7.4
bcrypt==4.1.1
python-multipart==0.0.6
requests==2.31.0
aiohttp==3.9.1
alembic==1.13.0
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2
email-validator==2.1.0
pandas==2.1.4
openai==1.3.0
